handle quiz submit for student

+++ database/migrations/2022_11_13_213632_create_take_quiz_table.php
@@ -20,7 +20,7 @@ class CreateTakeQuizTable extends Migration
             $table->dateTime('start_time');
             $table->dateTime('end_time');
             $table->dateTime('submit_time')->nullable();
-            $table->tinyInteger('score')->nullable();
+            $table->double('score')->nullable();
             $table->integer('total');
             $table->integer('number_correct')->nullable();
             $table->timestamps();
